Updated training scripts

+++ pushback/vexu_comp.py
@@ -21,13 +21,13 @@ class VexUCompGame(PushBackGame):
         if robots is None:
             robots = [
                 Robot(name="red_robot_0", team="red", size=RobotSize.INCH_24, 
-                      start_position=np.array([-42.0, 24.0])),
+                      start_position=np.array([-42.0, 24.0], dtype=np.float32)),
                 Robot(name="red_robot_1", team="red", size=RobotSize.INCH_15, 
-                      start_position=np.array([-46.5, -24.0])),
+                      start_position=np.array([-46.5, -24.0], dtype=np.float32)),
                 Robot(name="blue_robot_0", team="blue", size=RobotSize.INCH_24, 
-                      start_position=np.array([42.0, 24.0])),
+                      start_position=np.array([42.0, 24.0], dtype=np.float32)),
                 Robot(name="blue_robot_1", team="blue", size=RobotSize.INCH_15, 
-                      start_position=np.array([46.5, -24.0])),
+                      start_position=np.array([46.5, -24.0], dtype=np.float32)),
             ]
         super().__init__(robots)
